STY: optimize: rename `QuasiNewtonApprox` to `HessianUpdateStrategy`
